backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-restx==1.3.0
requests==2.31.0
boto3==1.34.0
openai==1.3.0
python-dotenv==1.0.0
werkzeug==2.3.7
PyJWT==2.8.0
pandasql==0.7.3
